backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python
torch
torchvision

# Optional: only needed for real MedGemma inference (api.py will run without it)
unsloth

# Testing
pytest
httpx
pytest-asyncio
